Title: 실수 입력받아 계산, Time: 135ms, Memory: 11MB, Status: Passed - CodeTree

+++ 240409/실수 입력받아 계산/calculation-by-inputting-a-real-number.java
@@ -3,8 +3,8 @@ public class Main {
     public static void main(String[] args) {
         // 여기에 코드를 작성해주세요.
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextDouble();
-        int b = sc.nextDouble();
-        System.out.printf("%.2f",(a+b));
+        double a = sc.nextDouble();
+        double b = sc.nextDouble();
+        System.out.printf("%.2f",a+b);
     }
 }
